Separate out pr comment controller from pr tree node
