+++ src/view/treeNodes/pullRequestNode.ts
@@ -4,52 +4,30 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as path from 'path';
-import { v4 as uuid } from 'uuid';
 import * as vscode from 'vscode';
-import { CommentHandler, registerCommentHandler, unregisterCommentHandler } from '../../commentHandlerResolver';
-import { DiffSide, IComment } from '../../common/comment';
 import { getCommentingRanges } from '../../common/commentingRanges';
 import { DiffChangeType, getModifiedContentFromDiffHunk, parseDiff } from '../../common/diffHunk';
 import { GitChangeType, SlimFileChange } from '../../common/file';
 import Logger from '../../common/logger';
 import { fromPRUri, toPRUri } from '../../common/uri';
-import { groupBy } from '../../common/utils';
 import { FolderRepositoryManager } from '../../github/folderRepositoryManager';
-import { GHPRComment, GHPRCommentThread, TemporaryComment } from '../../github/prComment';
-import { PullRequestModel, ReviewThreadChangeEvent } from '../../github/pullRequestModel';
-import {
-	CommentReactionHandler,
-	createVSCodeCommentThreadForReviewThread,
-	updateCommentReviewState,
-	updateCommentThreadLabel,
-	updateThread,
-} from '../../github/utils';
+import { PullRequestModel } from '../../github/pullRequestModel';
 import { getInMemPRContentProvider } from '../inMemPRContentProvider';
 import { DescriptionNode } from './descriptionNode';
 import { DirectoryTreeNode } from './directoryTreeNode';
 import { InMemFileChangeNode, RemoteFileChangeNode } from './fileChangeNode';
 import { TreeNode } from './treeNode';
 
-export class PRNode extends TreeNode implements CommentHandler, vscode.CommentingRangeProvider, CommentReactionHandler {
+export class PRNode extends TreeNode implements vscode.CommentingRangeProvider {
 	static ID = 'PRNode';
 
 	private _fileChanges: (RemoteFileChangeNode | InMemFileChangeNode)[] | undefined;
 	private _commentController?: vscode.CommentController;
-	public get commentController(): vscode.CommentController | undefined {
-		return this._commentController;
-	}
-
-	private _prCommentController?: vscode.Disposable & { commentThreadCache: { [key: string]: GHPRCommentThread[] } };
-	private _openPREditors: vscode.TextEditor[] = [];
 	private _disposables: vscode.Disposable[] = [];
 
 	private _inMemPRContentProvider?: vscode.Disposable;
 
 	private _command: vscode.Command;
-
-	private _commentHandlerId: string;
-	private _hasInitializedThreads: boolean = false;
-	private _pendingCommentThreadAdds: GHPRCommentThread[] = [];
 
 	public get command(): vscode.Command {
 		return this._command;
@@ -66,8 +44,6 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 		private _isLocal: boolean,
 	) {
 		super();
-		this._commentHandlerId = uuid();
-		registerCommentHandler(this._commentHandlerId, this);
 	}
 
 	// #region Tree
@@ -99,13 +75,11 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 				);
 			}
 
-			if (!this._prCommentController || !this._commentController) {
+			if (!this._commentController) {
 				await this.resolvePRCommentController();
 			}
 
-			await this.initializeThreadsInOpenEditors(vscode.window.visibleTextEditors);
 			await this.pullRequestModel.validateDraftMode();
-			await this.refreshContextKey(vscode.window.activeTextEditor);
 
 			const result: TreeNode[] = [descriptionNode];
 			const layout = vscode.workspace.getConfiguration('githubPullRequests').get<string>('fileListLayout');
@@ -133,23 +107,26 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 		}
 	}
 
-	private async resolvePRCommentController(): Promise<
-		vscode.Disposable & { commentThreadCache: { [key: string]: GHPRCommentThread[] } }
-	> {
-		if (this._prCommentController) {
-			return this._prCommentController;
-		}
-
+	private async resolvePRCommentController(): Promise<void> {
 		await this.pullRequestModel.githubRepository.ensureCommentsController();
 		this._commentController = this.pullRequestModel.githubRepository.commentsController!;
-		this._prCommentController = this.pullRequestModel.githubRepository.commentsHandler!.registerCommentController(
-			this.pullRequestModel.number,
-			this,
+
+		this._disposables.push(
+			this.pullRequestModel.githubRepository.commentsHandler!.registerCommentingRangeProvider(
+				this.pullRequestModel.number,
+				this,
+			),
+		);
+
+		this._disposables.push(
+			this.pullRequestModel.githubRepository.commentsHandler!.registerCommentController(
+				this.pullRequestModel.number,
+				this.pullRequestModel,
+				this._folderReposManager,
+			),
 		);
 
 		this.registerListeners();
-
-		return this._prCommentController;
 	}
 
 	private registerListeners(): void {
@@ -162,27 +139,6 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 						}
 					});
 				}
-
-				const commentThreadCache = (await this.resolvePRCommentController()).commentThreadCache;
-				for (const key in commentThreadCache) {
-					commentThreadCache[key].forEach(thread => {
-						updateCommentReviewState(thread, newDraftMode);
-					});
-				}
-			}),
-		);
-
-		this._disposables.push(
-			vscode.window.onDidChangeVisibleTextEditors(async e => {
-				this.onDidChangeOpenEditors(e);
-			}),
-		);
-
-		this._disposables.push(this.pullRequestModel.onDidChangeReviewThreads(e => this.onDidChangeReviewThreads(e)));
-
-		this._disposables.push(
-			vscode.window.onDidChangeActiveTextEditor(async e => {
-				await this.refreshContextKey(e);
 			}),
 		);
 	}
@@ -283,182 +239,6 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 		});
 	}
 
-	private getPREditors(editors: vscode.TextEditor[]): vscode.TextEditor[] {
-		return editors.filter(editor => {
-			if (editor.document.uri.scheme !== 'pr') {
-				return false;
-			}
-
-			const params = fromPRUri(editor.document.uri);
-
-			if (!params || params.prNumber !== this.pullRequestModel.number) {
-				return false;
-			}
-
-			return true;
-		});
-	}
-
-	private getCommentThreadCacheKey(fileName: string, isBase: boolean): string {
-		return `${fileName}-${isBase ? 'original' : 'modified'}`;
-	}
-
-	private addThreadsForEditors(
-		editors: vscode.TextEditor[],
-		commentThreadCache: { [key: string]: GHPRCommentThread[] },
-	): void {
-		const reviewThreads = this.pullRequestModel.reviewThreadsCache;
-		const threadsByPath = groupBy(reviewThreads, thread => thread.path);
-		editors.forEach(editor => {
-			const { fileName, isBase } = fromPRUri(editor.document.uri);
-			if (threadsByPath[fileName]) {
-				commentThreadCache[this.getCommentThreadCacheKey(fileName, isBase)] = threadsByPath[fileName]
-					.filter(
-						thread =>
-							(thread.diffSide === DiffSide.LEFT && isBase) ||
-							(thread.diffSide === DiffSide.RIGHT && !isBase),
-					)
-					.map(thread => {
-						const range = new vscode.Range(
-							new vscode.Position(thread.line - 1, 0),
-							new vscode.Position(thread.line - 1, 0),
-						);
-
-						return createVSCodeCommentThreadForReviewThread(
-							editor.document.uri,
-							range,
-							thread,
-							this._commentController,
-						);
-					});
-			}
-		});
-	}
-
-	private async initializeThreadsInOpenEditors(editors: vscode.TextEditor[]): Promise<void> {
-		if (this._hasInitializedThreads) {
-			return;
-		}
-
-		this._hasInitializedThreads = true;
-		const commentThreadCache = (await this.resolvePRCommentController()).commentThreadCache;
-
-		const prEditors = this.getPREditors(editors);
-		this._openPREditors = prEditors;
-		this.addThreadsForEditors(prEditors, commentThreadCache);
-	}
-
-	private onDidChangeOpenEditors(editors: vscode.TextEditor[]): void {
-		const commentThreadCache = this._prCommentController!.commentThreadCache;
-		const prEditors = this.getPREditors(editors);
-		const removed = this._openPREditors.filter(x => !prEditors.includes(x));
-		const added = prEditors.filter(x => !this._openPREditors.includes(x));
-		this._openPREditors = prEditors;
-
-		removed.forEach(editor => {
-			const { fileName, isBase } = fromPRUri(editor.document.uri);
-			const key = this.getCommentThreadCacheKey(fileName, isBase);
-			const threads = commentThreadCache[key] || [];
-			threads.forEach(t => t.dispose());
-			delete commentThreadCache[key];
-		});
-
-		if (added.length) {
-			this.addThreadsForEditors(added, commentThreadCache);
-		}
-	}
-
-	private onDidChangeReviewThreads(e: ReviewThreadChangeEvent): void {
-		const commentThreadCache = this._prCommentController.commentThreadCache;
-		e.added.forEach(thread => {
-			const fileName = thread.path;
-			const index = this._pendingCommentThreadAdds.findIndex(t => {
-				const samePath = this.gitRelativeRootPath(t.uri.path) === thread.path;
-				const sameLine = t.range.start.line + 1 === thread.line;
-				return samePath && sameLine;
-			});
-
-			let newThread: GHPRCommentThread;
-			if (index > -1) {
-				newThread = this._pendingCommentThreadAdds[index];
-				newThread.threadId = thread.id;
-				newThread.comments = thread.comments.map(c => new GHPRComment(c, newThread));
-				this._pendingCommentThreadAdds.splice(index, 1);
-			} else {
-				const openPREditors = this.getPREditors(vscode.window.visibleTextEditors);
-				const matchingEditor = openPREditors.find(editor => {
-					const query = fromPRUri(editor.document.uri);
-					const sameSide =
-						(thread.diffSide === DiffSide.RIGHT && !query.isBase) ||
-						(thread.diffSide === DiffSide.LEFT && query.isBase);
-					return query.fileName === fileName && sameSide;
-				});
-
-				if (matchingEditor) {
-					const range = new vscode.Range(
-						new vscode.Position(thread.line - 1, 0),
-						new vscode.Position(thread.line - 1, 0),
-					);
-
-					newThread = createVSCodeCommentThreadForReviewThread(
-						matchingEditor.document.uri,
-						range,
-						thread,
-						this._commentController,
-					);
-				}
-			}
-
-			const key = this.getCommentThreadCacheKey(thread.path, thread.diffSide === DiffSide.LEFT);
-			if (commentThreadCache[key]) {
-				commentThreadCache[key].push(newThread);
-			} else {
-				commentThreadCache[key] = [newThread];
-			}
-		});
-
-		e.changed.forEach(thread => {
-			const key = this.getCommentThreadCacheKey(thread.path, thread.diffSide === DiffSide.LEFT);
-			const index = commentThreadCache[key].findIndex(t => t.threadId === thread.id);
-			if (index > -1) {
-				const matchingThread = commentThreadCache[key][index];
-				updateThread(matchingThread, thread);
-			}
-		});
-
-		e.removed.forEach(async thread => {
-			const key = this.getCommentThreadCacheKey(thread.path, thread.diffSide === DiffSide.LEFT);
-			const index = commentThreadCache[key].findIndex(t => t.threadId === thread.id);
-			if (index > -1) {
-				const matchingThread = commentThreadCache[key][index];
-				commentThreadCache[key].splice(index, 1);
-				matchingThread.dispose();
-			}
-		});
-	}
-
-	private async refreshContextKey(editor: vscode.TextEditor | undefined) {
-		if (!editor) {
-			return;
-		}
-
-		const editorUri = editor.document.uri;
-		if (editorUri.scheme !== 'pr') {
-			return;
-		}
-
-		const params = fromPRUri(editorUri);
-		if (!params || params.prNumber !== this.pullRequestModel.number) {
-			return;
-		}
-
-		this.setContextKey(this.pullRequestModel.hasPendingReview);
-	}
-
-	private setContextKey(inDraftMode: boolean): void {
-		vscode.commands.executeCommand('setContext', 'prInDraft', inDraftMode);
-	}
-
 	getTreeItem(): vscode.TreeItem {
 		const currentBranchIsForThisPR = this.pullRequestModel.equals(this._folderReposManager.activePullRequest);
 
@@ -487,29 +267,6 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 		};
 	}
 
-	// #endregion
-
-	// #region Helper
-	hasCommentThread(thread: GHPRCommentThread): boolean {
-		if (thread.uri.scheme !== 'pr') {
-			return false;
-		}
-
-		const params = fromPRUri(thread.uri);
-
-		if (!params || params.prNumber !== this.pullRequestModel.number) {
-			return false;
-		}
-
-		return true;
-	}
-	private updateCommentThreadComments(thread: GHPRCommentThread, newComments: (GHPRComment | TemporaryComment)[]) {
-		thread.comments = newComments;
-		updateCommentThreadLabel(thread);
-	}
-
-	// #endregion
-
 	async provideCommentingRanges(
 		document: vscode.TextDocument,
 		_token: vscode.CancellationToken,
@@ -532,8 +289,6 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 
 		return undefined;
 	}
-
-	// #endregion
 
 	// #region Document Content Provider
 	private async provideDocumentContent(uri: vscode.Uri): Promise<string> {
@@ -649,235 +404,11 @@ export class PRNode extends TreeNode implements CommentHandler, vscode.Commentin
 		return '';
 	}
 
-	// #endregion
-
-	private getCommentSide(thread: GHPRCommentThread): DiffSide {
-		const query = fromPRUri(thread.uri);
-		return query.isBase ? DiffSide.LEFT : DiffSide.RIGHT;
-	}
-
-	// #region comment
-	public async createOrReplyComment(
-		thread: GHPRCommentThread,
-		input: string,
-		isSingleComment: boolean,
-		inDraft?: boolean,
-	): Promise<void> {
-		const hasExistingComments = thread.comments.length;
-		const isDraft = isSingleComment
-			? false
-			: inDraft !== undefined
-			? inDraft
-			: this.pullRequestModel.hasPendingReview;
-		const temporaryCommentId = this.optimisticallyAddComment(thread, input, isDraft);
-
-		try {
-			if (hasExistingComments) {
-				await this.reply(thread, input, isSingleComment);
-			} else {
-				const fileName = this.gitRelativeRootPath(thread.uri.path);
-				const side = this.getCommentSide(thread);
-				this._pendingCommentThreadAdds.push(thread);
-				await this.pullRequestModel.createReviewThread(input, fileName, thread.range.start.line + 1, side, isSingleComment);
-			}
-
-			if (isSingleComment) {
-				await this.pullRequestModel.submitReview();
-			}
-		} catch (e) {
-			vscode.window.showErrorMessage(`Creating comment failed: ${e}`);
-
-			thread.comments = thread.comments.map(c => {
-				if (c instanceof TemporaryComment && c.id === temporaryCommentId) {
-					c.mode = vscode.CommentMode.Editing;
-				}
-
-				return c;
-			});
-		}
-	}
-
-	private optimisticallyAddComment(thread: GHPRCommentThread, input: string, inDraft: boolean): number {
-		const currentUser = this._folderReposManager.getCurrentUser(this.pullRequestModel);
-		const comment = new TemporaryComment(thread, input, inDraft, currentUser);
-		this.updateCommentThreadComments(thread, [...thread.comments, comment]);
-		return comment.id;
-	}
-
-	private optimisticallyEditComment(thread: GHPRCommentThread, comment: GHPRComment): number {
-		const currentUser = this._folderReposManager.getCurrentUser(this.pullRequestModel);
-		const temporaryComment = new TemporaryComment(
-			thread,
-			comment.body instanceof vscode.MarkdownString ? comment.body.value : comment.body,
-			!!comment.label,
-			currentUser,
-			comment,
-		);
-		thread.comments = thread.comments.map(c => {
-			if (c instanceof GHPRComment && c.commentId === comment.commentId) {
-				return temporaryComment;
-			}
-
-			return c;
-		});
-
-		return temporaryComment.id;
-	}
-
-	private reply(thread: GHPRCommentThread, input: string, isSingleComment: boolean): Promise<IComment | undefined> {
-		const replyingTo = thread.comments[0];
-		if (replyingTo instanceof GHPRComment) {
-			return this.pullRequestModel.createCommentReply(input, replyingTo._rawComment.graphNodeId, isSingleComment);
-		} else {
-			// TODO can we do better?
-			throw new Error('Cannot respond to temporary comment');
-		}
-	}
-
-	public async editComment(thread: GHPRCommentThread, comment: GHPRComment | TemporaryComment): Promise<void> {
-		if (comment instanceof GHPRComment) {
-			const temporaryCommentId = this.optimisticallyEditComment(thread, comment);
-			try {
-				await this.pullRequestModel.editReviewComment(
-					comment._rawComment,
-					comment.body instanceof vscode.MarkdownString ? comment.body.value : comment.body,
-				);
-			} catch (e) {
-				vscode.window.showErrorMessage(`Editing comment failed ${e}`);
-
-				thread.comments = thread.comments.map(c => {
-					if (c instanceof TemporaryComment && c.id === temporaryCommentId) {
-						return new GHPRComment(comment._rawComment, thread);
-					}
-
-					return c;
-				});
-			}
-		} else {
-			this.createOrReplyComment(
-				thread,
-				comment.body instanceof vscode.MarkdownString ? comment.body.value : comment.body,
-				false,
-			);
-		}
-	}
-
-	public async deleteComment(thread: GHPRCommentThread, comment: GHPRComment | TemporaryComment): Promise<void> {
-		if (comment instanceof GHPRComment) {
-			await this.pullRequestModel.deleteReviewComment(comment.commentId);
-		} else {
-			thread.comments = thread.comments.filter(c => !(c instanceof TemporaryComment && c.id === comment.id));
-		}
-
-		await this.pullRequestModel.validateDraftMode();
-	}
-	// #endregion
-
-	private gitRelativeRootPath(comparePath: string) {
-		// get path relative to git root directory. Handles windows path by converting it to unix path.
-		return path.relative(this._folderReposManager.repository.rootUri.path, comparePath).replace(/\\/g, '/');
-	}
-
-	// #region Review
-	public async startReview(thread: GHPRCommentThread, input: string): Promise<void> {
-		const temporaryCommentId = this.optimisticallyAddComment(thread, input, true);
-
-		try {
-			const fileName = this.gitRelativeRootPath(thread.uri.path);
-			const side = this.getCommentSide(thread);
-			this._pendingCommentThreadAdds.push(thread);
-			await this.pullRequestModel.createReviewThread(input, fileName, thread.range.start.line + 1, side);
-
-			this.setContextKey(true);
-		} catch (e) {
-			vscode.window.showErrorMessage(`Starting a review failed: ${e}`);
-
-			thread.comments = thread.comments.map(c => {
-				if (c instanceof TemporaryComment && c.id === temporaryCommentId) {
-					c.mode = vscode.CommentMode.Editing;
-				}
-
-				return c;
-			});
-		}
-	}
-
-	public async finishReview(thread: GHPRCommentThread, input: string): Promise<void> {
-		try {
-			await this.createOrReplyComment(thread, input, false, false);
-			await this.pullRequestModel.submitReview();
-			this.setContextKey(false);
-		} catch (e) {
-			vscode.window.showErrorMessage(`Failed to submit the review: ${e}`);
-		}
-	}
-
-	public async deleteReview(): Promise<void> {
-		await this.pullRequestModel.deleteReview();
-		this.setContextKey(false);
-	}
-
-	private async createCommentOnResolve(thread: GHPRCommentThread, input: string): Promise<void> {
-		const pendingReviewId = await this.pullRequestModel.getPendingReviewId();
-		await this.createOrReplyComment(thread, input, !pendingReviewId);
-	}
-
-	public async resolveReviewThread(thread: GHPRCommentThread, input?: string): Promise<void> {
-		try {
-			if (input) {
-				await this.createCommentOnResolve(thread, input);
-			}
-
-			await this.pullRequestModel.resolveReviewThread(thread.threadId);
-		} catch (e) {
-			vscode.window.showErrorMessage(`Resolving conversation failed: ${e}`);
-		}
-	}
-
-	public async unresolveReviewThread(thread: GHPRCommentThread, input?: string): Promise<void> {
-		try {
-			if (input) {
-				await this.createCommentOnResolve(thread, input);
-			}
-
-			await this.pullRequestModel.unresolveReviewThread(thread.threadId);
-		} catch (e) {
-			vscode.window.showErrorMessage(`Unresolving conversation failed: ${e}`);
-		}
-	}
-
-	// #endregion
-
-	// #region Reaction
-	async toggleReaction(comment: GHPRComment, reaction: vscode.CommentReaction): Promise<void> {
-		if (comment.parent!.uri.scheme !== 'pr') {
-			return;
-		}
-
-		if (
-			comment.reactions &&
-			!comment.reactions.find(ret => ret.label === reaction.label && !!ret.authorHasReacted)
-		) {
-			// add reaction
-			await this.pullRequestModel.addCommentReaction(comment._rawComment.graphNodeId, reaction);
-		} else {
-			await this.pullRequestModel.deleteCommentReaction(comment._rawComment.graphNodeId, reaction);
-		}
-	}
-
-	// #endregion
-
 	dispose(): void {
 		super.dispose();
 
-		unregisterCommentHandler(this._commentHandlerId);
-
 		if (this._inMemPRContentProvider) {
 			this._inMemPRContentProvider.dispose();
-		}
-
-		if (this._prCommentController) {
-			this._prCommentController.dispose();
 		}
 
 		this._commentController = undefined;
